fix: dos bugs reales que el primer run del CI destapó
1. eslint: not found — eslint era dependencia transitiva (de @nuxt/eslint)
   y el layout estricto de pnpm no expone los binarios de transitivas en
   node_modules/.bin. Local funcionaba por deriva de un node_modules viejo.
   Ahora es dependencia directa: quien ejecuta un comando, lo declara.

2. round(double precision, integer) no existe en PostgreSQL — la
   verificación de la semilla (01_seed.sql) medía st_length()/1000 en
   double y PostgreSQL solo hace round(numeric, int). Cast explícito a
   numeric. Este SQL no se había ejecutado nunca contra una base real:
   el workflow que lo corre llevaba tres días sin poder subir.

Validado con instalación limpia + los 18 tests.

+++ sql/01_seed.sql
@@ -88,7 +88,7 @@ commit;
 
 -- ── Verificación inmediata de la semilla ──────────────────────────────────
 select id,
-       round(st_length(ruta) / 1000, 2) as km_de_ruta,
+       round((st_length(ruta) / 1000)::numeric, 2) as km_de_ruta,
        ventana,
        plazas
 from conductores
